Update DepositGui class to use TriumphGui
